Add a text editor decorator
This doesn't use any contribution point, just extension API.

The extension registers a hook to be called whenever the user
does a save while in a text editor. The hook gets the contents
of the file and looks for any occurences of the word "hello".
For all matches, it adds a decoration that colors the word
green with a solid white 2-pixel border.