Added a webworker that interacts with the Webview
The Webview now launches a Web Worker. It sends the worker
a message every second, and the worker responds immediately.
To see this all working, you need to view the console writes
by opening the Webview Developer Tools in the launched
VS code instance (via the command palette).

The web worker code is inlined in the HTML java script instead
of in a dedicated js file because of Theia issue:
   https://github.com/eclipse-theia/theia/issues/12828