Try (unsuccessfully) to introduce plugin API to backend
This commit shows that it's not possible to introduce custom plugin
API that's tied to the mainland backend. The Hello World
extension binds a second MainPluginApiProvider, this time from
the backend insteaed of the frontend. But it will never get
injected/exercised, as proven by the lack of this in stdout

   'HelloWorldBakendMainPluginApiProvider.initialize() called'

This is not surprising since the only code that references
MainPluginApiProvider is in Theia frontend modules. This commit
merely serves to prove it at runtime.

Introducing a new API namespace from a Theia extension involves
contributing a MainPluginApiProvider (via DI). The previous commit
shows that working well when the Theia extension's frontend is what's
supporting the API. But imagine the Theia extension having a backend
service that it wants to expose to Theia plugins. Currently there
is no direct way to do that, since MainPluginApiProvider is not
injected by any backend Theia code.