Demonstrate timing issue with activation events
If you invoke menu item
  Help > Say Hello
before the plugin-host starts up, you get a message in English.

If you launch the app and wait for the plugin-host to come up
then invoke the menu item, you'll get the message in Spanish.

This behavioral difference is the result of a deficiency in
HostedPluginSupport.activateByEvent(activationEvent: string).
I'll be opening an issue in the Theia github with more details.